+++ src/answer.py
@@ -15,22 +15,20 @@ from .page import Page
 
 
 class Answer(Page):
-	# FIXME! Try to get rid of the arguments fullmove_number and turn and
-	# derive them from the board instead.
-	# pylint: disable=too-many-arguments
-	def __init__(self, move: str, fullmove_number: int, turn: Color,
-	             board: Board, colour: Color) -> None:
+	def __init__(self, move: str, board: Board, colour: Color):
+		fullmove_number = board.fullmove_number
+		turn = not board.turn
 		self.move = move
 		self.fullmove_number = fullmove_number
 		Page.__init__(self, colour=colour, turn=turn)
 		self.set_board(board)
 
-	def render(self) -> str:
+	def render(self, note_id) -> str:
 		rendered = str(self.fullmove_number) + '.'
 		if not self.turn:
 			rendered += '...'
 		rendered += ' ' + self.move
-		rendered += self.extra_html()
+		rendered += self.extra_html(note_id)
 
 		return rendered