+++ src/dialog.py
@@ -8,87 +8,41 @@
 # http://www.wtfpl.net/ for more details.
 
 import os
+import html
 from pathlib import Path
-from typing import Dict, List
+import traceback
+from typing import List, Literal, Union
 
-from aqt import mw, AnkiQt
+from aqt import mw, AnkiQt, qconnect
 from aqt.operations import QueryOp
 # pylint: disable=no-name-in-module
-from aqt.qt import (QComboBox, QDialog,  # type: ignore[attr-defined]
-                    QDialogButtonBox, QFileDialog,  # type: ignore[attr-defined]
-                    QGridLayout, QLabel,  # type: ignore[attr-defined]
-                    QListWidget, QListWidgetItem,  # type: ignore[attr-defined]
-                    QPushButton, Qt) # type: ignore[attr-defined]
-from aqt.utils import showCritical, showInfo
+from aqt.qt import (QComboBox, QDialog, # type: ignore[attr-defined]
+                    QDialogButtonBox, QFileDialog, # type: ignore[attr-defined]
+                    QGridLayout, QLabel, # type: ignore[attr-defined]
+                    QListWidget, QListWidgetItem, # type: ignore[attr-defined]
+                    QPushButton, Qt, QMessageBox, # type: ignore[attr-defined]
+					QDesktopServices, QUrl) # type: ignore[attr-defined]
+from aqt.utils import showCritical, showInfo, showWarning
+from anki.utils import no_bundled_libs
+
+from config import Config
 
 from .importer import Importer
-
-
-class _Config():
-	# pylint: disable=too-few-public-methods
-	def __init__(self):
-		# pylint: disable=too-many-branches
-
-		if mw is not None:
-			config = mw.addonManager.getConfig(__name__)
-			if not config:
-				config = {}
-			col = mw.col
-
-			if ('colour' in config and isinstance(config['colour'], str)
-				and 'black' == config['colour']):
-				self.colour = 'black'
-			else:
-				self.colour = 'white'
-
-			if 'notetype' in config and isinstance(config['notetype'], str):
-				notetype = config['notetype']
-			else:
-				notetype = _('Basic')
-			if notetype in list(map(lambda m: m['name'], col.models.all())):
-				self.notetype = notetype
-			else:
-				self.notetype = str(None)
-
-			self.files: Dict[str, List[str]] = {'white': [], 'black': []}
-			if 'files' in config and isinstance(config['files'], dict):
-				files = config['files']
-				if 'white' in files and isinstance(config['files']['white'], list):
-					self.files['white'] = []
-					for filename in files['white']:
-						if isinstance(filename, str):
-							self.files['white'].append(filename)
-				if 'black' in files and isinstance(config['files']['black'], list):
-					self.files['black'] = []
-					for filename in files['black']:
-						if isinstance(filename, str):
-							self.files['black'].append(filename)
-
-			self.decks: Dict[str, str] = {}
-			if 'decks' in config and isinstance(config['decks'], dict):
-				decks = config['decks']
-				if ('white' in decks and isinstance(decks['white'], str)
-					and decks['white'] in list(
-						map(lambda d: d['name'], col.decks.all()))):
-					self.decks['white'] = decks['white']
-				else:
-					self.decks['white'] = ''
-				if ('black' in decks and isinstance(decks['black'], str)
-					and decks['black'] in list(
-						map(lambda d: d['name'], col.decks.all()))):
-					self.decks['black'] = decks['black']
-				else:
-					self.decks['black'] = ''
+from .config_reader import ConfigReader
 
 
 class ImportDialog(QDialog):
+
+
 	# pylint: disable=too-few-public-methods, too-many-instance-attributes
 	def __init__(self) -> None:
 		super().__init__()
 
-		self.config = _Config()
+		mw.col.media.trash_files(['foobar.svg'])
 
-		self.setWindowTitle(_('Import Opening PGN'))
+		self.config = ConfigReader().get_config()
+
+		self.setWindowTitle(_('Import PGN File'))
 
 		self.layout = QGridLayout()
 		self.setLayout(self.layout)
@@ -98,6 +52,7 @@ class ImportDialog(QDialog):
 		self.layout.addWidget(self.colour_combo, 0, 1)
 		self.colour_combo.addItem(_('White'))
 		self.colour_combo.addItem(_('Black'))
+		self.updating = False
 		self.colour_combo.currentIndexChanged.connect(self._colour_changed)
 
 		self.layout.addWidget(QLabel(_('Deck')), 1, 0)
@@ -107,8 +62,10 @@ class ImportDialog(QDialog):
 		if mw is not None:
 			for deck in mw.col.decks.all():
 				decknames.append(deck['name'])
+
 		for deckname in sorted(decknames):
 			self.deck_combo.addItem(deckname)
+		self.deck_combo.currentIndexChanged.connect(self._deck_changed)
 
 		self.layout.addWidget(QLabel(_('Input Files')), 2, 0)
 		self.file_list = QListWidget()
@@ -147,91 +104,198 @@ class ImportDialog(QDialog):
 		self._fill_dialog()
 
 	def _colour_changed(self) -> None:
+		if self.updating:
+			return
+		self.updating = True
+
 		if self.colour_combo.currentIndex() == 1:
 			colour = 'black'
 		else:
 			colour = 'white'
 		config = self.config
 
-		wanted = config.decks[colour]
-		for i in range(self.deck_combo.count()):
-			if wanted == self.deck_combo.itemText(i):
-				self.deck_combo.setCurrentIndex(i)
-				break
+		deck_id = config['decks'][colour]
+		deck = mw.col.decks.get(deck_id)
+		if deck is not None:
+			name = deck['name']
+			for i in range(self.deck_combo.count()):
+				if name == self.deck_combo.itemText(i):
+					self.deck_combo.setCurrentIndex(i)
+					break
+
+			self.file_list.clear()
+
+			if str(deck_id) in config['imports']:
+				record = config['imports'][str(deck_id)]
+				for filename in record['files']:
+					self.file_list.addItem(filename)
+
+		self.updating = False
+
+	def _deck_changed(self):
+		if self.updating:
+			return
+		self.updating = True
 
 		self.file_list.clear()
-		for filename in config.files[colour]:
-			self.file_list.addItem(filename)
 
-	def _fill_dialog(self) -> None:
-		config = self.config
-		colour = config.colour
+		deck_name = self.deck_combo.currentText()
+		deck_id = mw.col.decks.id_for_name(deck_name)
+		if deck_id is not None and str(deck_id) in self.config['imports']:
+			record = self.config['imports'][str(deck_id)]
+			self._set_colour_combo(record['colour'])
+
+			for filename in record['files']:
+				self.file_list.addItem(filename)
+
+		self.updating = False
+
+	def _set_colour_combo(self, colour: Literal['white', 'black']):
 		if 'black' == colour:
 			self.colour_combo.setCurrentIndex(1)
 		else:
 			self.colour_combo.setCurrentIndex(0)
 
-		if config.decks[colour] is not None:
-			wanted = config.decks[colour]
-			for i in range(self.deck_combo.count()):
-				if wanted == self.deck_combo.itemText(i):
-					self.deck_combo.setCurrentIndex(i)
-					break
+	def _fill_dialog(self) -> None:
+		config = self.config
+		colour = config['colour']
+		if colour is None:
+			colour = 'white'
 
-		if config.notetype is not None:
-			wanted = config.notetype
-			for i in range(self.model_combo.count()):
-				if wanted == self.model_combo.itemText(i):
-					self.model_combo.setCurrentIndex(i)
-					break
+		self._set_colour_combo(colour)
 
 		self.file_list.clear()
-		for filename in config.files[colour]:
-			self.file_list.addItem(filename)
+
+		if config['decks'][colour] is not None:
+			deck_id = config['decks'][colour]
+			deck = mw.col.decks.get(did=deck_id, default=False)
+
+			for i in range(self.deck_combo.count()):
+				if deck['name'] == self.deck_combo.itemText(i):
+					self.deck_combo.setCurrentIndex(i)
+
+			self.file_list.clear()
+			if str(deck_id) in config['imports']:
+				record = config['imports'][str(deck_id)]
+				for filename in record['files']:
+					self.file_list.addItem(filename)
+					break
+
+		if config['notetype'] is not None:
+			notetype_id = config['notetype']
+			name = mw.col.models.get(notetype_id)
+			if name is not None:
+				for i in range(self.model_combo.count()):
+					if name == self.model_combo.itemText(i):
+						self.model_combo.setCurrentIndex(i)
+						break
 
 	def accept(self) -> None:
 		assert isinstance(mw, AnkiQt)
 
-		def _on_success(counts: tuple[int, int, int, int, int]) -> None:
-			msgs = (
-			    ngettext('%d note inserted.', '%d notes inserted.',
-			             counts[0]) % (counts[0]),
-			    ngettext('%d note updated.', '%d notes updated.', counts[1]) %
-			    (counts[1]),
-			    ngettext('%d note deleted.', '%d notes deleted.', counts[2]) %
-			    (counts[2]),
-			    ngettext('%d image created.', '%d images created.',
-			             counts[3]) % (counts[3]),
-			    ngettext('%d image deleted.', '%d images deleted.',
-			             counts[4]) % (counts[4]),
-			)
+		def _on_success(counts: Union[Exception, tuple[int, int, int, int, int]]):
+			if isinstance(counts, Exception):
+				self._show_exception(counts)
+				return
+
+			if sum(counts) == 0:
+				msg = _('No changes since last import into this deck.')
+			else:
+				msgs = (
+					ngettext('%d note inserted.', '%d notes inserted.',
+							counts[0]) % (counts[0]),
+					ngettext('%d note updated.', '%d notes updated.', counts[1]) %
+					(counts[1]),
+					ngettext('%d note deleted.', '%d notes deleted.', counts[2]) %
+					(counts[2]),
+					ngettext('%d image created.', '%d images created.',
+							counts[3]) % (counts[3]),
+					ngettext('%d image deleted.', '%d images deleted.',
+							counts[4]) % (counts[4]),
+				)
+				msg = ' '.join(msgs)
+
 			mw.reset()
-			showInfo(' '.join(msgs))
+			mw.addonManager.writeConfig(__name__, self.config)
 
-		def _do_import(config, _unused) -> tuple[int, int, int, int, int]:
-			importer = Importer(
-			    collection=mw.col,
-			    filenames=config['files'][config['colour']],
-			    notetype=config['notetype'],
-			    colour=('white' == config['colour']),
-			    deck_name=config['decks'][config['colour']],
-			)
-			return importer.run()
+			showInfo(msg)
 
-		try:
-			if not self.file_list.count():
-				raise RuntimeError(_('No input files specified!'))
-			config = self.config.save(self) # type: ignore[attr-defined]
+		def _do_import(_) -> Union[Exception, tuple[int, int, int, int, int]]:
+			try:
+				colour = self.config['colour']
+				deck_id = self.config['decks'][colour]
+				record = self.config['imports'][str(deck_id)]
+				filenames = record['files']
+				notetype_id = self.config['notetype']
+
+				importer = Importer(
+					collection=mw.col,
+					deck_id=deck_id,
+					notetype_id=notetype_id,
+					filenames=filenames,
+					colour=('white' == colour),
+				)
+				return importer.run()
+			except Exception as e:
+				return e
+
+		if not self.file_list.count():
+			raise RuntimeError(_('No input files specified!'))
+		self.config = self._save_config()
+		if self.config is None:
+			self.reject()
+		else:
 			assert isinstance(mw, AnkiQt)
 			op = QueryOp(
-			    parent=mw,
-			    op=lambda _unused: _do_import(config, _unused),
-			    success=_on_success,
+				parent=mw,
+				op=lambda _unused: _do_import(_unused),
+				success=_on_success,
 			)
 			op.with_progress().run_in_background()
 			super().accept()
-		except OSError as e:
-			showCritical(str(e))
+
+	def _show_exception(self, e: Exception):
+		ftb = list(traceback.format_tb(e.__traceback__))
+
+		msgs = (
+			_('An error occurred!'),
+			_('Clicking the help button will open a web page explaining how to report a bug.'),
+			_('Please include the following information in your bug report:'),
+			'<hr />',
+			_('Exception type:') + ' ' + html.escape(type(e).__name__),
+			_('Exception message:') + ' ' + html.escape(str(e)),
+			_('Traceback:'),
+		)
+		msg = '<br />'.join(msgs) + '<br  />'.join(ftb)
+		print(msg)
+
+		parent = mw.app.activeWindow() or mw
+		msg_box = QMessageBox(parent)
+		msg_box.setIcon(QMessageBox.Icon.Critical)
+		msg_box.setTextFormat(Qt.TextFormat.MarkdownText)
+		msg_box.setText(msg)
+		msg_box.setWindowTitle(_('Chess Opening Trainer'))
+
+		# This is the reason, whey the customBtn argument for `showInfo()`
+		# cannot be used.  You can only call setDefault() or use
+		# qconnect() with the return value of addButton() which is a
+		# pointer to QPushButton.
+		ok_button = msg_box.addButton(QMessageBox.StandardButton.Ok)
+		ok_button.setDefault(True)
+
+		def open_link(link: str):
+			with no_bundled_libs():
+				QDesktopServices.openUrl(QUrl(link))
+
+		help_button = msg_box.addButton(QMessageBox.StandardButton.Help)
+		link = _('https://www.guido-flohr.net/practice-chess-openings-with-anki/#report-bugs')
+		help_button.clicked = qconnect(
+			help_button.clicked,
+			lambda: open_link(link),
+		)
+		help_button.setAutoDefault(False)
+
+		return msg_box.exec()
 
 	def _select_input_file(self) -> None:
 		if self.file_list.count():
@@ -250,29 +314,40 @@ class ImportDialog(QDialog):
 			self.file_list.addItems(
 			    [str(Path(filename)) for filename in filenames])
 
+	def _save_config(self) -> Config:
+		colour_index = self.colour_combo.currentIndex()
+		if colour_index == 1:
+			colour = 'black'
+		else:
+			colour = 'white'
 
-def _save_config(self, dlg: ImportDialog) -> dict:
-	colour_index = dlg.colour_combo.currentIndex()
-	if colour_index == 1:
-		colour = 'black'
-	else:
-		colour = 'white'
-	self.colour = colour
-	self.notetype = dlg.model_combo.currentText()
-	self.files[colour] = []
-	for i in range(dlg.file_list.count()):
-		self.files[colour].append(dlg.file_list.item(i).text())
-	self.decks[colour] = dlg.deck_combo.currentText()
-	config = {
-		'colour': self.colour,
-		'notetype': self.notetype,
-		'files': self.files,
-		'decks': self.decks,
-	}
+		col = mw.col
 
-	if mw is not None:
-		mw.addonManager.writeConfig(__name__, config)
+		deck_name = self.deck_combo.currentText()
+		deck_id = col.decks.id_for_name(deck_name)
+		if deck_id is None:
+			showWarning(_('The selected deck does not exist! Try again!'))
+			return None
 
-	return config
+		notetype_name = self.model_combo.currentText()
 
-_Config.save = _save_config # type: ignore[attr-defined]
+		notetype_id = col.models.id_for_name(notetype_name)
+		if notetype_id is None:
+			showWarning(_('The selected note type does not exist! Try again!'))
+			return None
+
+		files: List[str] = []
+		for i in range(self.file_list.count()):
+			files.append(self.file_list.item(i).text())
+
+		self.config['colour'] = colour
+		self.config['decks'][colour] = deck_id
+
+		self.config['imports'][str(deck_id)] = {
+			'colour': colour,
+			'files': files,
+		}
+
+		self.config['notetype'] = notetype_id
+
+		return self.config